started v1 notebooks
starting to implement path input

started catlaborator

basic elaborator working

compatible with catlog theory

rough implementation of quoting/pretty printing

This is so that elaboration errors can be more informative.

The idea for going from e-graph id to term syntax is that we store the
range of generators in each neutral notebook, so that we can search for
where a generator was originally synthesized and display a path to that
location.

This is somewhat janky, especially because if we ever have non-neutral
notebooks, then this won't make sense anymore. Specifically, this relies
on the fact that the generator ids are an inorder traversal of the tree
of notebooks.

There is still a bit to do, specifically pretty-printing non-generators
and pretty-printing whole notebooks.

fixed todos

tiling ui for multi-notebook

notebooks in tiling interface

ready to start calling into catlaborator from frontend

basic compositional notebooks working

fix tests