computad arbitrary for similar proptest as skelgraph and hashgraph
warning that not necessarily the correct number of vertices and edges as inputs num_v and num_e
demand
this is because the strategy for producing vertices or edges can produce repeats
so when making a vec out of them, not enough unique vertices/edges can end up occuring.

I don't think this is enough to merit putting in a prop_filter or prop_filter_map. Those inputs
are more what you'd call guidelines