BUG: Inconsistent graph visualization when node order is unstable.
This bug is not currently visible on `main` but surfaced during ongoing
work on #631. To see it, shuffle the list of judgments before calling
`modelToGraphviz`. The rendered SVG can then be inconsistent with the
computed Graphviz layout.

I believe this is a reactivity bug in Solid's `Suspense`. At least,
removing the `Suspense` seems to fix the problem.