BUG: Remove "tether" in notebook drag-and-drop to fix embed problem.
Despite `createEffect`'s claim that refs will be set before the effect
is run, when the notebook editor is embedded in Patchwork (a React app),
the ref `notebookRef` is undefined and an error is raised by
`dropTargetForElements`. My workaround is to simply remove this code,
which does not appear to change the drag-and-drop behavior.